feat: include CrimesInit
