--- src/modules/crimes/components/CrimeHeader/crimeHeader.scss
@@ -1,10 +1,0 @@
-@import '../../../../assets/scss/mixins';
-@import '../../../../assets/scss/variables';
-
-.crimeHeader {
-  align-items: center;
-  display: flex;
-  gap: 1.5rem;
-  justify-content: space-between;
-  margin: 0 0 1rem;
-}
