+++ src/modules/crimes/pages/Crimes/Crimes.tsx
@@ -1,26 +1,42 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import Button from '../../../../components/Button/Button';
+import Card from '../../../../components/Card/Card';
 import MainLayout from '../../../../components/MainLayout/MainLayout';
 import CrimeDetailModal from '../../components/CrimeDetailModal/CrimeDetailModal';
-import CrimeHeader from '../../components/CrimeHeader/CrimeHeader';
 import CrimeModal from '../../components/CrimeModal/CrimeModal';
 import CrimesTable from '../../components/CrimesTable/CrimesTable';
-import { ContextCrimeProvider } from '../../context/ContextCrime';
+import { ContextCrime } from '../../context/ContextCrime';
+import { Types } from '../../context/crime.reducer';
 import './crimes.scss';
 
 const Crimes: React.FC = () => {
+  const { dispatch } = useContext(ContextCrime);
+
+  const onOpen = () => {
+    dispatch({
+      type: Types.SetModal,
+      payload: {
+        active: true,
+        mode: 'add',
+        data: null,
+      },
+    });
+  };
+
   return (
-    <ContextCrimeProvider>
-      <MainLayout className="crimesPage">
-        <>
-          <CrimeModal />
-          <CrimeDetailModal />
-        </>
-        <>
-          <CrimeHeader />
-          <CrimesTable />
-        </>
-      </MainLayout>
-    </ContextCrimeProvider>
+    <MainLayout className="crimesPage">
+      <>
+        <CrimeModal />
+        <CrimeDetailModal />
+      </>
+      <Card className="crimesPage__header">
+        <h1>Lista de crímenes</h1>
+        <Button buttonType="secondary" onClick={onOpen}>
+          Registrar crímen
+        </Button>
+      </Card>
+      <CrimesTable />
+    </MainLayout>
   );
 };
