+++ src/modules/crimes/pages/Crimes/crimes.scss
@@ -2,4 +2,11 @@
 @import '../../../../assets/scss/variables';
 
 .crimesPage {
+  &__header {
+    align-items: center;
+    display: flex;
+    gap: 1.5rem;
+    justify-content: space-between;
+    margin: 0 0 1rem;
+  }
 }
